ESLint fixes
