+++ frontend/plan/src/components/search/old/filter.js
@@ -1,6 +1,6 @@
 import React from "react";
 import connect from "react-redux/es/connect/connect";
-import { OutClickable } from "../../old/dropdown";
+import { OutClickable } from "../../Dropdown";
 // import { toggleSearchFilterShown } from "../../../actions";
 
 class SearchFilter extends OutClickable {
